model_1/submodules/TimeSeriesDL/requirements/requirements.txt
absl-py==1.0.0
alembic==1.7.7
attrs==21.4.0
autopage==0.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
colorlog==6.6.0
cycler==0.11.0
fonttools==4.32.0
google-auth==2.6.5
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
kiwisolver==1.4.2
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
oauthlib==3.2.0
optuna==2.10.0
packaging==21.3
pbr==5.8.1
Pillow==9.1.0
prettytable==3.2.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pyperclip==1.8.2
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.35
stevedore==3.5.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing-extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.1
zipp==3.8.0
